separate search function into extra file and fix current UI integration
