+++ index.html
@@ -66,6 +66,7 @@
             ></ul>
         </main>
 
-        <script src="./main.js"></script>
+        <script type="module" src="main.js"></script>
+        <script type="module" src="search.js"></script>
     </body>
 </html>
