+++ journal.html
@@ -73,6 +73,7 @@
             ></ul>
         </main>
 
-        <script src="journal.js"></script>
+        <script type="module" src="search.js"></script>
+        <script type="module" src="journal.js"></script>
     </body>
 </html>
